backend/requirements.txt
asgiref==3.8.1
Django==4.2.20
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2

# Зависимости для работы в Docker
gunicorn==22.0.0
celery==5.4.0
redis==5.0.1
flower==2.0.1
whitenoise==6.6.0

# Добавляем недостающие зависимости и исправляем версии
# Для запуска на сервере
pytz==2023.3
autopep8==2.0.4
requests==2.31.0
urllib3>=2.0.0
vine==5.1.0
